Changed code to use a mixin
